:sparkles: Merge all tests in one file for DoublyLinkedList
